refactor(auto-pr): move onto the shared forge module
Fixes forge detection on self-hosted GitHub and GitLab. The plugin classified
remotes with remoteUrl.includes("github.com"), which is false for a host like
gitlab.mycorp.com, so detectForge returned null, shouldRun returned false, and
auto-PR silently did nothing on every enterprise install. The shared module
matches the parsed host instead, and falls back to a gh/glab probe for hosts
naming neither forge.
